refactor: Service 인터페이스 제거
